Replace size_t with int in Settings
This change was motivated by reading Insomniac Games' C++ style guide. Their reasoning makes sense to me!

The style guide: https://gist.github.com/Kerollmops/fcad27cfef9e3552cb75a3d201494ba6

+++ src/Settings.cpp
@@ -22,8 +22,8 @@ namespace MiniFPS {
         playerDistanceToProjectionPlane = 0;
     }
 
-    Settings::Settings(const std::string& version, size_t screenWidth, size_t screenHeight, float renderRayIncrement,
-                       size_t renderDistance, bool vSync,
+    Settings::Settings(const std::string& version, int screenWidth, int screenHeight, float renderRayIncrement,
+                       int renderDistance, bool vSync,
                        float fieldOfView, float speedModifier, float rotationModifier, float playerStartX,
                        float playerStartY, float playerStartAngle, float playerDistanceToProjectionPlane,
                        const std::string& levelPath, const std::map<std::string, std::string> &fontPaths) {
@@ -57,7 +57,7 @@ namespace MiniFPS {
         json settingsAsJson = json::parse(f);
         int major, minor, patch;
         std::string version;
-        size_t screenWidth, screenHeight, renderDistance;
+        int screenWidth, screenHeight, renderDistance;
         float renderRayIncrement, fieldOfView, speedModifier, rotationModifier, playerStartX, playerStartY,
                 playerStartAngle, playerDistanceToProjectionPlane;
         std::string levelPath;
